+++ internal/services/bucket_item.go
@@ -26,6 +26,8 @@ type IBucketItemService interface {
 	DeleteBucketItemsById(id []string) error
 	DeleteBucketItems(bucketUID string) error
 	DeleteBucketItemByKeyName(bucketUID string, key string) error
+	DeleteBucketItemsByKeyName(bucketUID string, keys []string) error
+	IncrementIntValue(bucketUID string, key string, amount int) error
 	FindBucketItemByID(id string) (*models.BucketItem, error)
 	FindBucketItemByKeyName(bucketUID string, key string) (*models.BucketItem, error)
 	ListBucketItems(bucketUID string) ([]models.BucketItem, error)
@@ -40,7 +42,8 @@ func NewBucketItemService(cfg *config.Config, bucketItemRepo repository.IBucketI
 }
 
 var (
-	ErrKeyIsEmpty = errors.New("key cannot be empty")
+	ErrKeyIsEmpty          = errors.New("bucket item key cannot be empty")
+	ErrBucketItemIDIsEmpty = errors.New("bucket item id cannot be empty")
 )
 
 // Creates a new bucket item
@@ -50,6 +53,9 @@ func (b *BucketItemService) CreateBucketItem(data dto.CreateBucketItemInputDTO, 
 	// validation
 	if utils.IsStringEmpty(data.Key) {
 		return &dto.CreateBucketItemOutputDTO{}, ErrKeyIsEmpty
+	}
+	if utils.IsStringEmpty(bucketUID) {
+		return &dto.CreateBucketItemOutputDTO{}, ErrBucketUIDIsEmpty
 	}
 	// check if the bucket UID exists
 	bucket, err := b.BucketRepository.FindBucketByUID(bucketUID)
@@ -71,6 +77,7 @@ func (b *BucketItemService) CreateBucketItem(data dto.CreateBucketItemInputDTO, 
 		BucketID:  bucket.ID,
 		Key:       data.Key,
 		Data:      data.Data,
+		TTL:       data.TTL,
 	}
 	newBucketItem.CreatedAt = primitive.NewDateTimeFromTime(time.Now())
 	newBucketItem.UpdatedAt = primitive.NewDateTimeFromTime(time.Now())
@@ -83,6 +90,7 @@ func (b *BucketItemService) CreateBucketItem(data dto.CreateBucketItemInputDTO, 
 		BucketUID: bucketUID,
 		Key:       data.Key,
 		Data:      data.Data,
+		TTL:       data.TTL,
 		CreatedAt: newBucketItem.CreatedAt,
 	}, nil
 }
@@ -91,12 +99,36 @@ func (b *BucketItemService) CreateBucketItem(data dto.CreateBucketItemInputDTO, 
 // Accepts the update data, bucket UID and key
 // Returns an error
 func (b *BucketItemService) UpdateBucketItemByKeyName(data dto.UpdateBucketItemInputDTO, bucketUID string, key string) error {
+	if utils.IsStringEmpty(bucketUID) {
+		return ErrBucketUIDIsEmpty
+	}
+	if utils.IsStringEmpty(key) {
+		return ErrKeyIsEmpty
+	}
 	updatedBucketItem := &models.BucketItem{
 		BucketUID: bucketUID,
 		Key:       data.Key,
 		Data:      data.Data,
+		TTL:       data.TTL,
 	}
 	err := b.BucketItemRepository.UpdateBucketItem(updatedBucketItem, key)
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+// Increments/decrements a key value by an amount
+// Accepts the bucket UID, item key, and amount
+// Returns an error
+func (b *BucketItemService) IncrementIntValue(bucketUID string, key string, amount int) error {
+	if utils.IsStringEmpty(bucketUID) {
+		return ErrBucketUIDIsEmpty
+	}
+	if utils.IsStringEmpty(key) {
+		return ErrKeyIsEmpty
+	}
+	err := b.BucketItemRepository.IncrementIntItem(bucketUID, key, amount)
 	if err != nil {
 		return err
 	}
@@ -107,6 +139,9 @@ func (b *BucketItemService) UpdateBucketItemByKeyName(data dto.UpdateBucketItemI
 // Accepts the string value of the bucket item object ID
 // Returns the found bucket item and an error
 func (b *BucketItemService) FindBucketItemByID(id string) (*models.BucketItem, error) {
+	if utils.IsStringEmpty(id) {
+		return &models.BucketItem{}, ErrBucketItemIDIsEmpty
+	}
 	bucketItem, err := b.BucketItemRepository.FindBucketItemByID(id)
 	if err != nil {
 		return &models.BucketItem{}, err
@@ -118,6 +153,12 @@ func (b *BucketItemService) FindBucketItemByID(id string) (*models.BucketItem, e
 // Accepts the bucket UID and key name values
 // Returns the found bucket item and an error
 func (b *BucketItemService) FindBucketItemByKeyName(bucketUID string, key string) (*models.BucketItem, error) {
+	if utils.IsStringEmpty(bucketUID) {
+		return &models.BucketItem{}, ErrBucketUIDIsEmpty
+	}
+	if utils.IsStringEmpty(key) {
+		return &models.BucketItem{}, ErrKeyIsEmpty
+	}
 	bucketItem, err := b.BucketItemRepository.FindBucketItemByKeyName(bucketUID, key)
 	if err != nil {
 		return &models.BucketItem{}, err
@@ -129,6 +170,9 @@ func (b *BucketItemService) FindBucketItemByKeyName(bucketUID string, key string
 // Accepts the bucket UID
 // Returns the found bucket items and an error
 func (b *BucketItemService) ListBucketItems(bucketUID string) ([]models.BucketItem, error) {
+	if utils.IsStringEmpty(bucketUID) {
+		return []models.BucketItem{}, ErrBucketUIDIsEmpty
+	}
 	bucketItems, err := b.BucketItemRepository.FindBucketItems(bucketUID)
 	if err != nil {
 		return []models.BucketItem{}, err
@@ -140,6 +184,9 @@ func (b *BucketItemService) ListBucketItems(bucketUID string) ([]models.BucketIt
 // Accepts the string value of the bucket item's object ID
 // Returns an error
 func (b *BucketItemService) DeleteBucketItemById(id string) error {
+	if utils.IsStringEmpty(id) {
+		return ErrBucketItemIDIsEmpty
+	}
 	err := b.BucketItemRepository.DeleteBucketItemById(id)
 	if err != nil {
 		return err
@@ -162,9 +209,31 @@ func (b *BucketItemService) DeleteBucketItemsById(ids []string) error {
 // Accepts the bucket UID and key name for the desired bucket item
 // Returns an error
 func (b *BucketItemService) DeleteBucketItemByKeyName(bucketUID string, key string) error {
+	if utils.IsStringEmpty(bucketUID) {
+		return ErrBucketUIDIsEmpty
+	}
+	if utils.IsStringEmpty(key) {
+		return ErrKeyIsEmpty
+	}
 	err := b.BucketItemRepository.DeleteBucketItemByKeyName(bucketUID, key)
 	if err != nil {
 		return err
+	}
+	return nil
+}
+
+// Delete bucket items by key name
+// Accepts the bucket UID and key names for the desired bucket items
+// Returns an error
+func (b *BucketItemService) DeleteBucketItemsByKeyName(bucketUID string, keys []string) error {
+	if utils.IsStringEmpty(bucketUID) {
+		return ErrBucketUIDIsEmpty
+	}
+	for _, key := range keys {
+		err := b.BucketItemRepository.DeleteBucketItemByKeyName(bucketUID, key)
+		if err != nil {
+			return err
+		}
 	}
 	return nil
 }
@@ -173,6 +242,9 @@ func (b *BucketItemService) DeleteBucketItemByKeyName(bucketUID string, key stri
 // Accepts the bucket UID of the bucket
 // Returns an error
 func (b *BucketItemService) DeleteBucketItems(bucketUID string) error {
+	if utils.IsStringEmpty(bucketUID) {
+		return ErrBucketUIDIsEmpty
+	}
 	err := b.BucketItemRepository.DeleteBucketItems(bucketUID)
 	if err != nil {
 		return err
